[H5Group] Initialise group prop if it exists
Before it was assumed that if the group property is None that the
underlying underlying HDF5 group did not exist, since instantiating a
group sets the property in such cases. With the new Container class, a
long living H5Group object but be created that does not have a group
prop set but it ends up being created in another place (presumably, from
a _create_obj() method).