[File] Header format string type fix
This has been tested to be crosscompatible across all combinations of
backends and python versions (2, 3.6, hdf5, h5py).

I've determined that the best way to handle the string type of format is
to set is to encode it to ascii in the property setter and for the
getter to return the string as it is read from the hdf5 file directly.

✈