[h5dataset] Catch H5Py indexing ValueError
H5Py throws ValueError for out-of-bounds indexing. We catch it and throw
IndexError instead, reusing the error message from H5Py.

This also makes indexing more flexible. Slices now work the same as in
numpy, so no slice is ever out-of-bounds. The corresponding test has
been updated accordingly.

This is an updated version of PR #303, which was a fix for #299.