[dimensions] Bugfix: Set .index properly on __init__
DimensionContainer failed to set the dimension properly when
initialising a Dimension that was being loaded from the file. It was
looking for an "index" attribute and was returning None. Should be using
the name of the HDF5 Group object.

Added tests to catch regressions.

Wrapping index in int() when returning and when setting the object
attribute to catch unwanted behaviour.