Run compatibility tests only when requested
The compatibility tests are no longer run by default when NIX is
detected. They can be run by specifying --nix-compat. If specified, the
tests are compiled and any errors will cause failure of the
compatibility tests. If the option is not specified, the tests are
skipped, even if NIX is available.

One reason for this change is that the compatibility tests were always
compiled when NIX was available, even when they are unselected, and
compilation slows down the tests.

With this change, they are only compiled and run when explicitly
enabled.