[Copy] Introduce the copy_x funcs to file and block Add copy func for entities between files
It is recursive.

[copy] Fix bug for copying mt, tag, da, df h5Group

Now the copied object will goes not directly under the block but blk/class_name/

[copy] Fix bug for object indexing after copying

The creation order in get_by_pos was not reachable in the original method. Now it is ok and mt, da, df and tags
can be reached by indexing after copying

Clean up

Change copy_object to append_copy

now the entity being appended to have the method instead of the one being
copied

[copy] Add sanity checks to the copy function

[copy] Divide the single copy function to several func for specific obj