Skip consistency type check for numpy data in Props
If data are given as numpy values, the check fails, even when the data
types are correct, for cases where the original data are (for example)
float32 and a numpy.float32 array is also given.  The type checker
infers the largest type variant for the individual values (float64) and
determines that the types are incompatible.

Fix is to let numpy arrays through without checking consistency, since
arrays are already consistent.