Upgrade script: Add file ID and update file version
This is a WIP and still needs to take care of old style Property
objects.

The script first checks the file header version to see if it's older
than the current. If it is, it continues to collect tasks that need to
be performed to complete the upgrade.

Tasks are created using functions that check if a change is required and
if so, they return a closure bound to the specific filename. The
returned function is then run after the user confirms they want to apply
the changes. If --force is specified, the changes are applied without
prompting.

All checks and changes are made in standalone functions that open the
file using h5py and close it immediately after. This makes each change
atomic (or atomic-ish) and reduces the chances that a file will be left
in an invalid state. There's still a chance the user will stop the
procedure after a change has been applied and before the file version in
the header is updated. This will put the file in a sort of invalid state
but rerunning the upgrade script should finalise the changes.

Regardless of the possibilities of corruption or invalid states, the
user is warned to make sure their files and data are backed up.