Remove header property setters
Header properties shouldn't have setters.  The format and version
attributes should only ever be added on creation.  A property setter
should never be exposed, since it communicates that the user can edit
the attributes freely, which they shouldn't since it would break the
file header check.

The version attribute is not overwritten if it exists when the method is
called.  This avoids resetting the version by calling the method on an
old file with a new library, which can be done since "private" methods
aren't really private.  The same check is not made for format since it
never changes.

Removed timestamp updating from the setter methods since it's called in
the constructor, which is only run on new file creation.