Upgrade script: Update properties to new format
Extra attributes, if set, become new properties to avoid data loss.
Uncertainties are added to the new uncertainty attribute of the
property, but only if there is one unique uncertainty for all values in
the old property.  If it is not unique and each value has its own
uncertainty, they are also added as an extra property.