Upgrade script: Recheck property before fixing
If you supply the same file twice (perhaps through a symlink), the task
collection procedure will put the same file on the list twice. This wont
be a problem with the file ID and version update, but the property
reformatting might cause issues.The script will crash when attempting to
get the "uncertainty" component of the property's value, so it shouldn't
cause any corruption.

Regardless, checking the property format again right before performing
the change will prevent the issue.