Upgrade script: Conditionally add and run tasks
Task preparation functions are responsible for returning the closure or
None if it's not needed.  The task is not added if it's not needed.

Even if it is added, the task checks if it needs to make the chance
right before making it (guards against multiple runs on the same file
after task collection).