[h5group] Don't delete empty containers
The initial consideration was to have "clean" deep objects that didn't
have lingering empty containers (like a Source container), but this
becomes a problem when the child object of a non-Container object is
removed (e.g., RangeDimension.ticks).  The result was the dimension
object itself was being deleted because it had no children between the
deletion of the ticks and the creation of the link.  We could reverse
the operation to get around this, but I think it's better to avoid this
risky self-deletion in general.