Fix DataFrame read tests for strings->bytes
String data and attributes are generally returned as strings.  H5Py
returns string data from a file as bytes and these are decoded to
strings before returning to the user.  Unfortunately, there's no clean
way to do the same with compound data types that contain string
elements.  Strings in DataFrame data are returned as bytes and tests are
adjusted accordingly.