[DataFrame] Swap __getitem__ and read_rows
Swapped the implementation of __getitem__ and read_rows as this makes
life a lot easier.
For one, __getitem__ takes care to give us slices unless the index is a
single integer value, which  saves us the trouble of having multiple
data retrieval paths.
We move the multi- vs single-row handling code to the conversion method
which checks if the data that was retrieved prior to converting is a
single row (ndarray with shape == ()).
The conversion method is only called if the dtype has "fields", which
indicates a compound/struct type.