[DataView] Use data object's _read_data()
Instead of reading from the underlying DataSet object's _read_data(),
call the linked data object's method.  In the case of a DataArray, this
will apply polynomial coefficients and the expansion origin to the data
slice before returning.

Test added for retrieving data with coefficients and expansion origin
through a MultiTag.

Fixes #482.