Use SliceMode in Tag.tagged_data()
The tagged_data() method allows defining the SliceMode which affects how
the right (stop) endpoint behaves with the 'stop_rule' argument.  This
affects the value that's returned when the final position (pos+ext)
matches exactly a tick on the corresponding dimension:
If Exclusive, the previous position is returned.
If Inclusive, the matching position is returned.

The index_of() methods of the Dimensions take a boolean `leq` which
defines whether to return the exact match when found (less or equal), or
the previous (strictly less).