test_doc_examples.py: initialize examples_path.
While trying to package the Python nixio module in Debian, most tests
from test_doc_examples.py are failing with the following error:

	    def handle_lif(self):
	>       lif_path = Path.joinpath(self.examples_path, "lif.py")
	E       AttributeError: 'TestDocumentationExamples' object has no attribute 'examples_path'

This looks to be caused by missing initialisation of the examples_path
attribute of the TestDocumentationExamples class.  Initialising the
attribute with the path to the documentation from the root of the
sources corrects the issue.

Signed-off-by: Étienne Mollier <emollier@debian.org>